feat(server): Refactorización a ServerDetailsContainer y ServerDetailsInfo
