--- app/(home)/servers/[id]/ServerDetails.types.ts
@@ -1,3 +1,0 @@
-export interface ServerDetailsProps {
-  id: string;
-}
