+++ app/(home)/servers/components/ServerDetailsContainer/index.ts
@@ -1,0 +1,1 @@
+export * from "./ServerDetailsContainer";
