+++ app/(home)/servers/components/ServerDetailsInfo/index.ts
@@ -1,0 +1,1 @@
+export * from './ServerDetailsInfo';
